[seller] refactor(230): file-upload-input을 widgets에서 shared/ui로 이동
