+++ apps/seller/src/features/register/account-verification/account-verification.ui.tsx
@@ -12,8 +12,8 @@ import {
   verifyAccount,
 } from '@/entity/register'
 import { extractServerMessage } from '@/shared/utils/extract-server-message'
-import { FileUploadInput } from '@/widgets/file-upload-input'
-import { InputField } from '@/widgets/input-field'
+import { FileUploadInput } from '../../../shared/ui/file-upload-input'
+import { InputField } from '../../../widgets/input-field'
 
 import { handleRegisterFileUpload, sanitizeAccountPayload } from '../lib'
 import { REGISTER_TOAST_MESSAGES } from '../register.constant'
